pedagogical-flow-extractor/requirements.txt
# ── Python dependencies 
# Install with: pip install -r requirements.txt
# ffmpeg must be installed separately before running the pipeline:


# Audio transcription (OpenAI Whisper)
openai-whisper==20250625

# Deep learning backend required by Whisper
torch==2.10.0
torchvision
torchaudio

# YouTube audio downloader
yt-dlp==2026.3.3

# Gemini API client
google-genai==1.66.0

# Graph construction and analysis
networkx==3.4.2

# Interactive graph visualisation (generates HTML output)
pyvis==0.3.2

# Loads GEMINI_API_KEY from a .env file
python-dotenv>=1.0.0
